[neptune/#169] Block editing Verified Field Visits & allow deleting Assessments/Maintenance records Enable deleting Assessments from the Field Visit Detail Page and ensure correct behavior when do
