+++ Source/Neptune.Web/Neptune.Web.csproj
@@ -294,6 +294,7 @@
     <Compile Include="Models\Generated\WaterQualityManagementPlanStatus.Binding.cs" />
     <Compile Include="Models\Generated\WaterQualityManagementPlanStatusPrimaryKey.Binding.cs" />
     <Compile Include="Views\FieldVisit\AssessmentDetail.cs" />
+    <Compile Include="Views\FieldVisit\AssessmentDetailViewData.cs" />
     <Compile Include="Views\FieldVisit\Detail.cs" />
     <Compile Include="Views\FieldVisit\DetailViewData.cs" />
     <Compile Include="Models\HydrologicSubarea.cs" />
