+++ Source/Neptune.Web/Views/FieldVisit/AssessmentDetail.cs
@@ -22,11 +22,6 @@ using System.Web.Mvc;
 using LtInfo.Common;
 using LtInfo.Common.HtmlHelperExtensions;
 using LtInfo.Common.Mvc;
-using Neptune.Web.Common;
-using Neptune.Web.Controllers;
-using Neptune.Web.Models;
-using Neptune.Web.Views.Shared;
-using Neptune.Web.Views.TreatmentBMPAssessment;
 
 namespace Neptune.Web.Views.FieldVisit
 {
@@ -35,31 +30,6 @@ namespace Neptune.Web.Views.FieldVisit
         public static void RenderPartialView(HtmlHelper html, AssessmentDetailViewData viewData)
         {
             html.RenderRazorSitkaPartial<AssessmentDetail, AssessmentDetailViewData>(viewData);
-        }
-    }
-
-    public class AssessmentDetailViewData
-    {
-        public Models.TreatmentBMPAssessment TreatmentBMPAssessment { get; }
-        public bool CurrentPersonCanManage { get; }
-        public bool CanEdit { get; }
-        public ScoreDetailViewData ScoreDetailViewData { get; }
-        public string EditBenchmarkAndThresholdUrl { get; }
-        public ImageCarouselViewData ImageCarouselViewData { get; }
-
-        public AssessmentDetailViewData(Person currentPerson, Models.TreatmentBMPAssessment treatmentBMPAssessment)
-        {
-            TreatmentBMPAssessment = treatmentBMPAssessment;
-            CurrentPersonCanManage = currentPerson.IsAssignedToStormwaterJurisdiction(treatmentBMPAssessment.TreatmentBMP.StormwaterJurisdiction);
-            ScoreDetailViewData = new ScoreDetailViewData(treatmentBMPAssessment);
-            EditBenchmarkAndThresholdUrl =
-                SitkaRoute<TreatmentBMPBenchmarkAndThresholdController>.BuildUrlFromExpression(x =>
-                    x.Instructions(treatmentBMPAssessment.TreatmentBMP));
-
-            CanEdit = CurrentPersonCanManage && treatmentBMPAssessment.CanEdit(currentPerson) && !treatmentBMPAssessment.IsAssessmentComplete();
-
-            var carouselImages = TreatmentBMPAssessment.TreatmentBMPAssessmentPhotos;
-            ImageCarouselViewData = new ImageCarouselViewData(carouselImages, 400);
         }
     }
 }
