+++ Source/Neptune.Web/Views/FieldVisit/DetailViewData.cs
@@ -19,6 +19,8 @@ namespace Neptune.Web.Views.FieldVisit
         public AssessmentDetailViewData InitialAssessmentViewData { get; }
         public AssessmentDetailViewData PostMaintenanceAssessmentViewData { get; }
         public Models.MaintenanceRecord MaintenanceRecord { get; }
+        public bool UserCanDeleteInitialAssessment { get; }
+        public bool UserCanDeletePostMaintenanceAssessment { get; }
 
         public DetailViewData(Person currentPerson, StormwaterBreadCrumbEntity stormwaterBreadCrumbEntity,
             Models.FieldVisit fieldVisit, AssessmentDetailViewData initialAssessmentViewData, AssessmentDetailViewData postMaintenanceAssessmentViewData) : base(currentPerson, stormwaterBreadCrumbEntity)
@@ -32,6 +34,14 @@ namespace Neptune.Web.Views.FieldVisit
             SubEntityName = FieldVisit.TreatmentBMP.TreatmentBMPName ?? "Preview Treatment BMP Field Visit";
             SubEntityUrl = FieldVisit.TreatmentBMP?.GetDetailUrl() ?? "#";
             PageTitle = FieldVisit.VisitDate.ToStringDate();
+            UserCanDeleteInitialAssessment = FieldVisit.InitialAssessment != null &&
+                                             new TreatmentBMPAssessmentManageFeature()
+                                                 .HasPermission(currentPerson, FieldVisit.InitialAssessment)
+                                                 .HasPermission;
+            UserCanDeletePostMaintenanceAssessment = FieldVisit.PostMaintenanceAssessment != null &&
+                                             new TreatmentBMPAssessmentManageFeature()
+                                                 .HasPermission(currentPerson, FieldVisit.PostMaintenanceAssessment)
+                                                 .HasPermission;
             UserCanDeleteMaintenanceRecord = FieldVisit.MaintenanceRecord != null &&
                                              new MaintenanceRecordManageFeature()
                                                  .HasPermission(currentPerson, FieldVisit.MaintenanceRecord)
